Changed Crop and Lock to Dark Mode
The only reason I made this fork was to have a dark mode version of the
Crop and Lock module.
I tried for hours to make sure I could keep both light and dark themes,
but C++ simply wouldn't let me.
I am not a C++ programmer, so I don't know how to fix this.

I now understood why no one bothered to make a dark mode version.

It turns out where and how to make a dark mode is relatively easy.
But making it in a decent way is a lot harder, after solving my problem
I spent a few hours trying to make it work in a way that could be sent
upstream to the main branch. I wasn't able to do that, so I decided to
make a fork of the Crop and Lock module.